refactor(#48): 도메인 정리에 의한 패키지 이동
