+++ src/main/java/com/najasin/domain/manual/controller/MyManualController.java
@@ -17,17 +17,17 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.najasin.domain.character.dto.param.CharacterItemsParam;
 import com.najasin.domain.character.service.CharacterService;
-import com.najasin.domain.manual.answer.service.AnswerService;
+import com.najasin.domain.manual.service.AnswerService;
 import com.najasin.domain.manual.dto.param.JffMyKeyword;
 import com.najasin.domain.manual.dto.request.MyManualCreateRequest;
 import com.najasin.domain.manual.dto.response.JffMyManualCreateResponse;
 import com.najasin.domain.manual.dto.response.JffMyManualResponse;
 import com.najasin.domain.manual.dto.param.JffMyQuestion;
-import com.najasin.domain.manual.keyword.entity.Keyword;
-import com.najasin.domain.manual.keyword.service.KeywordService;
-import com.najasin.domain.manual.question.entity.Question;
-import com.najasin.domain.manual.question.service.QuestionService;
-import com.najasin.domain.manual.userKeyword.service.UserKeywordService;
+import com.najasin.domain.manual.entity.keyword.Keyword;
+import com.najasin.domain.manual.service.KeywordService;
+import com.najasin.domain.manual.entity.question.Question;
+import com.najasin.domain.manual.service.QuestionService;
+import com.najasin.domain.manual.service.UserKeywordService;
 import com.najasin.domain.user.entity.User;
 import com.najasin.domain.user.entity.userType.UserUserType;
 import com.najasin.domain.user.service.UserUserTypeService;
