+++ src/main/java/com/najasin/domain/manual/dto/param/JffMyAnswer.java
@@ -1,7 +1,7 @@
 package com.najasin.domain.manual.dto.param;
 
-import com.najasin.domain.manual.answer.entity.Answer;
-import com.najasin.domain.manual.question.entity.Question;
+import com.najasin.domain.manual.entity.answer.Answer;
+import com.najasin.domain.manual.entity.question.Question;
 import com.najasin.domain.user.entity.User;
 
 public record JffMyAnswer(Long id,
