+++ src/main/java/com/najasin/domain/manual/entity/answer/AnswerId.java
@@ -1,6 +1,6 @@
-package com.najasin.domain.comment.entity;
+package com.najasin.domain.manual.entity.answer;
 
-import com.najasin.domain.manual.question.entity.Question;
+import com.najasin.domain.manual.entity.question.Question;
 import com.najasin.domain.user.entity.User;
 import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
@@ -9,12 +9,13 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
+
 @Embeddable
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CommentId implements Serializable {
+public class AnswerId implements Serializable {
     private User user;
     private Question question;
 }
