+++ src/main/java/com/najasin/domain/manual/entity/comment/CommentId.java
@@ -1,6 +1,6 @@
-package com.najasin.domain.manual.answer.entity;
+package com.najasin.domain.manual.entity.comment;
 
-import com.najasin.domain.manual.question.entity.Question;
+import com.najasin.domain.manual.entity.question.Question;
 import com.najasin.domain.user.entity.User;
 import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
@@ -9,13 +9,12 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
-
 @Embeddable
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class AnswerId implements Serializable {
+public class CommentId implements Serializable {
     private User user;
     private Question question;
 }
