+++ src/main/java/com/najasin/domain/manual/repository/KeywordRepository.java
@@ -1,6 +1,6 @@
-package com.najasin.domain.manual.keyword.repository;
+package com.najasin.domain.manual.repository;
 
-import com.najasin.domain.manual.keyword.entity.Keyword;
+import com.najasin.domain.manual.entity.keyword.Keyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface KeywordRepository extends JpaRepository<Keyword, Long> {
