+++ src/main/java/com/najasin/domain/manual/repository/UserKeywordRepository.java
@@ -1,9 +1,9 @@
-package com.najasin.domain.manual.userKeyword.repository;
+package com.najasin.domain.manual.repository;
 
 import java.util.List;
 
-import com.najasin.domain.manual.userKeyword.entity.UserKeyword;
-import com.najasin.domain.manual.userKeyword.entity.UserKeywordId;
+import com.najasin.domain.manual.entity.userKeyword.UserKeyword;
+import com.najasin.domain.manual.entity.userKeyword.UserKeywordId;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserKeywordRepository extends JpaRepository<UserKeyword, UserKeywordId> {
