+++ src/main/java/com/najasin/domain/manual/service/KeywordService.java
@@ -1,4 +1,4 @@
-package com.najasin.domain.manual.keyword.service;
+package com.najasin.domain.manual.service;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.najasin.domain.manual.dto.param.JffMyKeyword;
-import com.najasin.domain.manual.keyword.entity.Keyword;
-import com.najasin.domain.manual.keyword.repository.KeywordRepository;
+import com.najasin.domain.manual.entity.keyword.Keyword;
+import com.najasin.domain.manual.repository.KeywordRepository;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
